+++ Amara/src/amara/game/maps/Map_Destroforest.java
@@ -146,7 +146,7 @@ public class Map_Destroforest extends Map{
             enemy.setComponent(new DirectionComponent(direction));
             enemy.setComponent(new AutoAggroComponent(15));
             enemy.setComponent(new CastSpellOnCooldownWhileAttackingComponent(0));
-            enemy.setComponent(new SetNewTargetSpellsOnCooldownComponent(new int[]{0}, new int[]{6}));
+            enemy.setComponent(new SetNewTargetSpellsOnCooldownComponent(new int[]{0}, new float[]{6}));
             enemy.setComponent(new TeamComponent(0));
             enemy.setComponent(new CampComponent(campEnemies2.getId(), position, direction));
         }
